Semana16Hackaton/jacbarreto/requirements.txt
django
gunicorn
django-heroku
djangorestframework
